Move BPM type back to BpmInput.tsx
